Add UI test infrastructure for compiler diagnostics
Extract shared test runner code from rue-spec into rue-test-runner
crate, then create rue-ui-tests for testing compiler warnings and
error messages. Unlike spec tests which verify language conformance,
UI tests verify the quality and content of diagnostic messages.

Initial test cases cover unreachable code and unused variable warnings.
test.sh now runs both test suites.

Closes tree2-n9t
